chore(web): fix lint errors (#2529)
* chore(web): fix lint errors

* f

* f

* f

* f

* f

* f

* f
