+++ web/src/contexts/SettingsContext.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useState, useEffect } from 'react';
+import { createContext, useContext } from 'react';
 
 export interface Settings {
   themeColor: string;
@@ -9,44 +9,7 @@ interface SettingsContextType {
   updateSettings: (newSettings: Partial<Settings>) => void;
 }
 
-const defaultSettings: Settings = {
-  themeColor: '#316DE6',
-};
-
-const SETTINGS_KEY = 'app-settings';
-
 export const SettingsContext = createContext<SettingsContextType | undefined>(undefined);
-
-export const SettingsProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-  const [settings, setSettings] = useState<Settings>(() => {
-    const saved = localStorage.getItem(SETTINGS_KEY);
-    const config = saved ? JSON.parse(saved) : defaultSettings;
-    if (!config.themeColor) {
-      config.themeColor = defaultSettings.themeColor;
-    }
-    return config;
-  });
-
-  useEffect(() => {
-    localStorage.setItem(SETTINGS_KEY, JSON.stringify(settings));
-  }, [settings]);
-
-  const updateSettings = (newSettings: Partial<Settings>) => {
-    setSettings((prev) => {
-      const updated = { ...prev, ...newSettings };
-      if (!updated.themeColor) {
-        updated.themeColor = defaultSettings.themeColor;
-      }
-      return updated;
-    });
-  };
-
-  return (
-    <SettingsContext.Provider value={{ settings, updateSettings }}>
-      {children}
-    </SettingsContext.Provider>
-  );
-};
 
 export const useSettings = (): SettingsContextType => {
   const context = useContext(SettingsContext);
@@ -54,4 +17,4 @@ export const useSettings = (): SettingsContextType => {
     throw new Error('useSettings must be used within a SettingsProvider');
   }
   return context;
-}; 
+};
